build tags for dev/prod
