+++ .github/workflows/build.yml
@@ -23,6 +23,8 @@ jobs:
         run: cp deps/wintun/bin/amd64/wintun.dll ./
       - name: build
         uses: goreleaser/goreleaser-action@v5
+        env:
+          BUILD_ENV: "development"
         with:
           version: latest
           args: build --clean --snapshot
