+++ .goreleaser.yml
@@ -20,6 +20,8 @@ builds:
       - -X github.com/superfly/flyctl/internal/buildinfo.buildDate={{.CommitDate}}
       - -X github.com/superfly/flyctl/internal/buildinfo.buildVersion={{.Version}}
       - -X github.com/superfly/flyctl/internal/buildinfo.commit={{.ShortCommit}}
+    flags:
+      - -tags={{ envOrDefault "BUILD_ENV" "production" }}
 
   - id: windows
     env:
@@ -34,6 +36,8 @@ builds:
       - -X github.com/superfly/flyctl/internal/buildinfo.buildDate={{.CommitDate}}
       - -X github.com/superfly/flyctl/internal/buildinfo.buildVersion={{.Version}}
       - -X github.com/superfly/flyctl/internal/buildinfo.commit={{.ShortCommit}}
+    flags:
+      - -tags={{ envOrDefault "BUILD_ENV" "production" }}
 
 archives:
   - id: windows
