+++ internal/buildinfo/env_production.go
@@ -1,0 +1,21 @@
+//go:build production
+
+package buildinfo
+
+import (
+	"time"
+
+	"github.com/superfly/flyctl/internal/version"
+)
+
+var environment = "production"
+
+func loadBuildTime() (err error) {
+	cachedBuildTime, err = time.Parse(time.RFC3339, buildDate)
+	return
+}
+
+func loadVersion() (err error) {
+	cachedVersion, err = version.Parse(buildVersion)
+	return
+}
